projetos-pessoais/telegram-bot/requirements.txt
python-telegram-bot==20.1
requests==2.28.2